Improvements for hdf
